+++ liblzma-rs/src/common/filter_flags_decoder.rs
@@ -51,7 +51,7 @@ pub struct lzma_filter {
 }
 pub const __DARWIN_NULL: *mut c_void = ::core::ptr::null_mut::<c_void>();
 pub const NULL: *mut c_void = __DARWIN_NULL;
-pub const LZMA_FILTER_RESERVED_START: c_ulonglong = (1 as c_ulonglong) << 62 as c_int;
+pub const LZMA_FILTER_RESERVED_START: c_ulonglong = 1 << 62;
 #[no_mangle]
 pub unsafe extern "C" fn lzma_filter_flags_decode(
     mut filter: *mut lzma_filter,
@@ -68,7 +68,7 @@ pub unsafe extern "C" fn lzma_filter_flags_decode(
         in_pos,
         in_size,
     ) as lzma_ret;
-    if ret_ as c_uint != LZMA_OK as c_uint {
+    if ret_ != LZMA_OK {
         return ret_;
     }
     if (*filter).id >= LZMA_FILTER_RESERVED_START as lzma_vli {
@@ -82,7 +82,7 @@ pub unsafe extern "C" fn lzma_filter_flags_decode(
         in_pos,
         in_size,
     ) as lzma_ret;
-    if ret__0 as c_uint != LZMA_OK as c_uint {
+    if ret__0 != LZMA_OK {
         return ret__0;
     }
     if (in_size.wrapping_sub(*in_pos) as lzma_vli) < props_size {
